+++ src/main/java/JdukeException.java
@@ -1,0 +1,5 @@
+public class JdukeException extends Exception {
+    public JdukeException(String message) {
+        super(message);
+    }
+}
